feat(theme): Add Forest Green theme and refactor theme system
This introduces a robust Enum-based theme system, adds the Green color palette, and updates settings to allow theme selection. Closes #1
