+++ app/src/main/java/com/example/librehabit/model/Theme.kt
@@ -1,0 +1,13 @@
+package com.example.librehabit.model
+
+enum class AppTheme(val displayName: String) {
+    PURPLE("Purple"),
+    FOREST_GREEN("Forest Green"),
+    SYSTEM_DYNAMIC("System Dynamic")
+}
+
+enum class DarkModePreference(val displayName: String) {
+    SYSTEM("System"),
+    LIGHT("Light"),
+    DARK("Dark")
+}
