+++ app/src/main/java/com/example/librehabit/ui/settings/SettingsScreen.kt
@@ -8,17 +8,7 @@ import android.net.Uri
 import android.widget.Toast
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -27,26 +17,8 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.automirrored.filled.OpenInNew
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -54,12 +26,16 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.example.librehabit.UnitSystem
+import com.example.librehabit.model.AppTheme
+import com.example.librehabit.model.DarkModePreference
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SettingsScreen(
-    isDarkMode: Boolean?,
-    onDarkModeChange: (Boolean?) -> Unit,
+    appTheme: AppTheme,
+    onAppThemeChange: (AppTheme) -> Unit,
+    darkModePreference: DarkModePreference,
+    onDarkModePreferenceChange: (DarkModePreference) -> Unit,
     unitSystem: UnitSystem,
     onUnitSystemChange: (UnitSystem) -> Unit,
     height: Float,
@@ -70,7 +46,7 @@ fun SettingsScreen(
     appVersion: String,
     onNavigateUp: () -> Unit
 ) {
-    var heightInput by remember(height) { mutableStateOf(height.toString()) }
+    var heightInput by remember(height) { mutableStateOf(if (height > 0) height.toString() else "") }
     val context = LocalContext.current
     val scrollState = rememberScrollState()
 
@@ -119,34 +95,40 @@ fun SettingsScreen(
                 .padding(16.dp)
                 .verticalScroll(scrollState)
         ) {
-            Text("Theme", style = MaterialTheme.typography.titleLarge)
+            Text("Appearance", style = MaterialTheme.typography.titleLarge)
+            Spacer(modifier = Modifier.height(8.dp))
+            Text("Dark Mode", style = MaterialTheme.typography.titleMedium)
             Column {
-                ThemeRadioButton(
-                    text = "Light",
-                    selected = isDarkMode == false,
-                    onClick = { onDarkModeChange(false) }
-                )
-                ThemeRadioButton(
-                    text = "Dark",
-                    selected = isDarkMode == true,
-                    onClick = { onDarkModeChange(true) }
-                )
-                ThemeRadioButton(
-                    text = "System",
-                    selected = isDarkMode == null,
-                    onClick = { onDarkModeChange(null) }
-                )
+                DarkModePreference.values().forEach { preference ->
+                    ThemeRadioButton(
+                        text = preference.displayName,
+                        selected = darkModePreference == preference,
+                        onClick = { onDarkModePreferenceChange(preference) }
+                    )
+                }
             }
+            Spacer(modifier = Modifier.height(16.dp))
+            Text("Color Theme", style = MaterialTheme.typography.titleMedium)
+            Column {
+                AppTheme.values().forEach { theme ->
+                    ThemeRadioButton(
+                        text = theme.displayName,
+                        selected = appTheme == theme,
+                        onClick = { onAppThemeChange(theme) }
+                    )
+                }
+            }
+
             Spacer(modifier = Modifier.height(24.dp))
             Text("Units", style = MaterialTheme.typography.titleLarge)
             Column {
                 UnitRadioButton(
-                    text = "Metric (kg)",
+                    text = "Metric (kg, cm)",
                     selected = unitSystem == UnitSystem.METRIC,
                     onClick = { onUnitSystemChange(UnitSystem.METRIC) }
                 )
                 UnitRadioButton(
-                    text = "Imperial (lbs)",
+                    text = "Imperial (lbs, in)",
                     selected = unitSystem == UnitSystem.IMPERIAL,
                     onClick = { onUnitSystemChange(UnitSystem.IMPERIAL) }
                 )
