+++ app/src/main/java/com/example/librehabit/ui/theme/Green.kt
@@ -1,0 +1,11 @@
+package com.example.librehabit.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Green80 = Color(0xFFB5CCB3)
+val GreenGrey80 = Color(0xFFC1C9BF)
+val LightGreen80 = Color(0xFFDDE6DA)
+
+val Green40 = Color(0xFF4C664A)
+val GreenGrey40 = Color(0xFF5A6359)
+val LightGreen40 = Color(0xFF727B70)
